+++ NEWS.md
@@ -4,7 +4,10 @@
 
 * A major rewrite/refactor of the underlying code that runs `tune_grid()`. This was an upgrade to add postprocessing and to modernize our parallel processing infrastructure. 
 
-* The pattern of `.config` values has changed from `Preprocessor{num}_Model{num}` to `pre{num}_mod{num}_post{num}`. The numbers include a zero when that element was static. For example, a value of `pre0_mod3_post4` means no preprocessors were tuned and the model and postprocessor(s) had at least three and four candidates, respectively.   
+* The pattern of `.config` values has changed. 
+
+   - For grid search, it changes from `Preprocessor{num}_Model{num}` to `pre{num}_mod{num}_post{num}`. The numbers include a zero when that element was static. For example, a value of `pre0_mod3_post4` means no preprocessors were tuned and the model and postprocessor(s) had at least three and four candidates, respectively. 
+   - For iterative search, the pattern is not `iter{num}` instead of `Iter{num}` and the numbers are now zero padded to sort better. For example, if there between 10 and 99 iterations, the first `.config` value is now `iter01` instead of `Iter1`.
 
 * The package will now log a backtrace for errors and warnings that occur during tuning. When a tuning process encounters issues, see the new `trace` column in the `collect_notes(.Last.tune.result)` output to find precisely where the error occurred (#873).